refactor testing utils
