+++ precompiles/pallet-democracy/src/tests.rs
@@ -31,7 +31,7 @@ use pallet_democracy::{
 	PreimageStatus, Vote, VoteThreshold, Voting,
 };
 use pallet_evm::{Call as EvmCall, Event as EvmEvent};
-use precompile_utils::{prelude::*, solidity, testing::*};
+use precompile_utils::{prelude::*, testing::*};
 use sp_core::{H160, H256, U256};
 use std::{convert::TryInto, str::from_utf8};
